apps/telephony_app/pyproject.toml
[tool.poetry]
name = "telephony-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pyngrok = "^7.1.6"
python-dotenv = "^1.0.1"
azure-cognitiveservices-speech = "^1.38.0"
vocode = "0.1.113"
vonage = "^3.16.0"
libhoney = "^2.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
